character pick up/drop off placements
each character is placed on the map above the stop where they will be picked up and dropped off. a red square next to a character represents where a character will be dropped off and a green square is where a character will be picked up. I will upload a final map without the squares and characters